% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_raster.R
\name{export_raster}
\alias{export_raster}
\title{Write raster}
\usage{
export_raster(x, file, FUN = raster::writeRaster,
  data_source = flyio_get_datasource(),
  bucket = flyio_get_bucket(data_source), dir = flyio_get_dir(),
  delete_file = TRUE, show_progress = FALSE, ...)
}
\arguments{
\item{x}{variable name}

\item{file}{path of the file to be written to}

\item{FUN}{the function using which the file is to write}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{dir}{the directory to store intermediate files}

\item{delete_file}{logical. to delete the file to be uploaded}

\item{show_progress}{logical. Shows progress of the upload operation.}

\item{...}{other parameters for the FUN function defined above}
}
\value{
No output
}
\description{
Write raster
}
\examples{
\dontrun{
flyio_set_datasource("gcs")
flyio_set_bucket("your-bucket-name")
r1 <- raster(nrows=108, ncols=21, xmn=0, xmx=10)
export_raster(r1, "raster-cloud.tif", writeRaster, format = "GTiff", dir = tempdir())
}
}
