% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_bucket.R
\name{flyio_get_bucket}
\alias{flyio_get_bucket}
\title{Get global bucket name for flyio}
\usage{
flyio_get_bucket(data_source = flyio_get_datasource())
}
\arguments{
\item{data_source}{the data source used for I/O. Default chooses the data source set using flyio_set_datasource()}
}
\value{
the string - bucket name stored
}
\description{
Get global bucket name to be used for all the functions in flyio
}
\details{
if the data source is local, then an empty string is returned
}
\examples{
# first setting the bucket for a data source
flyio_set_bucket(bucket = "socialcops-test", data_source = "S3")
# retrieving the bucket for S3
flyio_get_bucket(data_source = "S3")
}
