% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_dir.R
\name{flyio_get_dir}
\alias{flyio_get_dir}
\title{Get global bucket name for flyio}
\usage{
flyio_get_dir()
}
\value{
the string - directory name
}
\description{
Get global directory where flyio functions will download intermidiate files
}
\details{
if the directory is not set using flyio_set_dir(), it will return the paste0(tempdir(),"/flyio")
}
\examples{
flyio_get_dir()
}
