% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_dir.R
\name{flyio_list_dir}
\alias{flyio_list_dir}
\title{List files in flyio tmp folder}
\usage{
flyio_list_dir()
}
\value{
the string - file names
}
\description{
Get the list of files downloaded by flyio in the default tmp folder
}
\examples{
flyio_list_dir()
}
