% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_dir.R
\name{flyio_remove_dir}
\alias{flyio_remove_dir}
\title{Delete files in flyio tmp folder}
\usage{
flyio_remove_dir()
}
\value{
files deleted
}
\description{
Delete the list of files downloaded by flyio in the default tmp folder
}
\examples{
flyio_remove_dir()
}
