% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_bucket.R
\name{flyio_set_bucket}
\alias{flyio_set_bucket}
\title{Set global bucket name for flyio}
\usage{
flyio_set_bucket(bucket, data_source = flyio_get_datasource())
}
\arguments{
\item{bucket}{the bucket name to be set}

\item{data_source}{the data source used for I/O. Default chooses the data source set using flyio_set_datasource()}
}
\value{
stores the bucket name in a global environment under flyioBucketGcs or flyioBucketS3
}
\description{
Set global bucket name to be used for all the functions in flyio
}
\examples{
flyio_set_bucket(bucket = "your-bucket-name", data_source = "S3")
}
