% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_dir.R
\name{flyio_set_dir}
\alias{flyio_set_dir}
\title{Set global directory for flyio to store data}
\usage{
flyio_set_dir(dir = paste0(tempdir(), "/flyio", Sys.getpid()))
}
\arguments{
\item{dir}{the directory to store intermediate files}
}
\value{
stores the directory in a global environment under CLOUD_DIR
}
\description{
Set global directory where flyio functions will download intermidiate files
}
\examples{
flyio_set_dir(dir = tempdir())
}
