% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bucket.R
\name{list_bucket}
\alias{list_bucket}
\title{List buckets for cloud storage}
\usage{
list_bucket(data_source = flyio_get_datasource(), gcs_project = "")
}
\arguments{
\item{data_source}{default to local. Possible options : gcs, s3, local. Case insensitive}

\item{gcs_project}{Project containing buckets to list in Google Cloud Storage}
}
\value{
vector of bucket names
}
\description{
List buckets for cloud storage
}
\examples{
# No buckets if data source is local
list_bucket(data_source = "local")
\dontrun{
flyio_set_datasource("s3")
flyio_auth() # authentication needed for S3
list_bucket()
}
}
