% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioFileDownload.R
\name{rioFileDownload}
\alias{rioFileDownload}
\title{Download file from cloud to local system}
\usage{
rioFileDownload(bucketpath, localfile, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source), overwrite = TRUE, ...)
}
\arguments{
\item{bucketpath}{path of file in the bucket}

\item{localfile}{path where the file needs to be downloaded. The filename and extension also needs to be present, if not the current file name will be considered}

\item{data_source}{the name of the data source, if not set globally}

\item{bucket}{the name of the bucket, if not set globally}

\item{overwrite}{logical. If the files should be overwritten if already present}

\item{...}{other parameters for gcs_get_object or save_object}
}
\value{
logical. if the operation was successful or not
}
\description{
Save a single file from the cloud to your local drive
}
\examples{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioFileDownload("mtcars.csv", "~/Downloads/dsada.csv", overwrite = T)
}
