% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioFileExists.R
\name{rioFileExists}
\alias{rioFileExists}
\title{Check if a file exists}
\usage{
rioFileExists(path, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source))
}
\arguments{
\item{path}{the entire path for the file}

\item{data_source}{the name of the data source, gcs, s3 or local; if not set globally}

\item{bucket}{the name of the bucket, if not set globally}
}
\value{
logical. if the file exists or not
}
\description{
Check if a file exists
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioFileExists(path = "tests/mtcars.csv")
}
}
