% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioFileUpload.R
\name{rioFileUpload}
\alias{rioFileUpload}
\title{Upload a file from the local system to cloud}
\usage{
rioFileUpload(localfile, bucketpath, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source), ...)
}
\arguments{
\item{localfile}{path of the file to be uploaded}

\item{bucketpath}{path there the file needs to be uploaded, the filename can or cannot be present}

\item{data_source}{the name of the data source, if not set globally. gcs or s3}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for gcs_upload}
}
\value{
the filename and path of the file in the bucket
}
\description{
Write a local file to the cloud, S3 or GCS
}
\examples{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioFileUpload("~/Downloads/dsada.csv", "tests/mtcars.csv")

}
