% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioFolderUpload.R
\name{rioFolderUpload}
\alias{rioFolderUpload}
\title{Upload a file from the local system to cloud}
\usage{
rioFolderUpload(localfolder, bucketpath, pattern = "*",
  overwrite = TRUE, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source), ...)
}
\arguments{
\item{localfolder}{path of the folder in which all the files are to be uploaded}

\item{bucketpath}{path of the folder in which the files are to be uploaded}

\item{data_source}{the name of the data source, if not set globally. can be gcs or s3}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for gcs/s3 upload}
}
\value{
the filename and path of the file in the bucket
}
\description{
Write a local file to the cloud, S3 or GCS
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioFileUpload("~/Downloads/", "tests/")
}
}
