% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioBucket.R
\name{rioGetBucket}
\alias{rioGetBucket}
\title{Get global bucket name for rIO}
\usage{
rioGetBucket(data_source = rioGetDataSource())
}
\arguments{
\item{data_source}{the data source used for IO. Default chooses the data source set using rioSetDataSource()}
}
\value{
the string - bucket name stored
}
\description{
Get global bucket name to be used for all the function in rIO
}
\details{
if the data source is local, then an empty string is returned
}
\examples{
# first setting the bucket for a data source
rioSetBucket(bucket = "socialcops-test", data_source = "S3")
# retrieving the bucket for S3
rioGetBucket(data_source = "S3")
}
