% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioDataSource.R
\name{rioGetDataSource}
\alias{rioGetDataSource}
\title{Get global Data Source name for rIO}
\usage{
rioGetDataSource()
}
\value{
the string - DataSource name stored
}
\description{
Get global data source name to be used for all the function in rIO. Returns the value stored using rioSetDataSource
}
\examples{
rioGetDataSource()
}
