% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioListFiles.R
\name{rioListFiles}
\alias{rioListFiles}
\title{List the Files in a Directory/Folder}
\usage{
rioListFiles(path = "", pattern = NULL, recursive = FALSE,
  ignore.case = FALSE, full.names = TRUE,
  data_source = rioGetDataSource(), bucket = rioGetBucket(data_source))
}
\arguments{
\item{path}{the folder for which the files needs to be listed}

\item{pattern}{an optional regular expression. Only file path names which match the regular expression will be returned.}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{ignore.case}{logical. Should pattern-matching be case-insensitive?}

\item{full.names}{logical. Should the entire path be returned or only after the path inputed.}

\item{data_source}{the name of the data source, gcs, s3 or local; if not set globally}

\item{bucket}{the name of the bucket, if not set globally}
}
\value{
a vector of full file names
}
\description{
list the files in cloud or locally - similar to list.files()
}
\examples{
\dontrun{
rioSetDataSource("s3")
rioSetBucket("socialcops-test")
gcsListFiles(path = "tests/", pattern = ".*csv")
}
}
