% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioReadRda.R
\name{rioReadRda}
\alias{rioReadRda}
\title{Read RDA file}
\usage{
rioReadRda(file, FUN = load, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source), ...)
}
\arguments{
\item{file}{path of the file to be read}

\item{FUN}{the function using which the file is to be read}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
the output of the FUN function
}
\description{
Read RData or rda file from anywhere
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
load("tests/googletest.rda")
}
}
