% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioReadShp.R
\name{rioReadShp}
\alias{rioReadShp}
\title{Read shapefiles}
\usage{
rioReadShp(dsn, layer, FUN = rgdal::readOGR, dsnlayerbind = F,
  data_source = rioGetDataSource(), bucket = rioGetBucket(data_source),
  ...)
}
\arguments{
\item{dsn}{path of the file to be read}

\item{layer}{the name of the shapefile without extension}

\item{FUN}{the function using which the file is to be read}

\item{dsnlayerbind}{if the FUN needs dsn and layer binded or not}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
the output of the FUN function
}
\description{
Read shapefiles data from anywhere using a function defined by you
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
t = rioReadShp("tests/shptest/", "testshp", FUN = readOGR, dsnlayerbind = F)
t = rioReadShp("tests/shptest/", "testshp", FUN = shapefile, dsnlayerbind = T)
}
}
