% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioReadTable.R
\name{rioReadTable}
\alias{rioReadTable}
\title{Read csv, Excel files, txt}
\usage{
rioReadTable(file, FUN = read.csv, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source), ...)
}
\arguments{
\item{file}{path of the file to be read}

\item{FUN}{the function using which the file is to be read}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
the output of the FUN function
}
\description{
Read tabular data from anywhere using a function defined by you
}
\examples{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioReadTable("tests/googletest.xlsx", read_excel)

}
