% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioBucket.R
\name{rioSetBucket}
\alias{rioSetBucket}
\title{Set global bucket name for rIO}
\usage{
rioSetBucket(bucket, data_source = rioGetDataSource())
}
\arguments{
\item{bucket}{the bucket name to be set}

\item{data_source}{the data source used for IO. Default chooses the data source set using rioSetDataSource()}
}
\value{
stores the bucket name in a global environment under rioBucketGcs or rioBucketS3
}
\description{
Set global bucket name to be used for all the function in rIO
}
\examples{
rioSetBucket(bucket = "socialcops-test", data_source = "S3")
}
