% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioDataSource.R
\name{rioSetDataSource}
\alias{rioSetDataSource}
\title{Set global DataSource name for rIO}
\usage{
rioSetDataSource(data_source)
}
\arguments{
\item{data_source}{the DataSource name to be set}
}
\value{
stores the DataSource name in a global environment under rioDataSource
}
\description{
Set global DataSource name to be used for all the function in rIO
}
\examples{
rioSetDataSource("local")
}
