% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioWriteRaster.R
\name{rioWriteRaster}
\alias{rioWriteRaster}
\title{Write raster to GCS}
\usage{
rioWriteRaster(x, file, FUN = raster::writeRaster,
  data_source = rioGetDataSource(), bucket = rioGetBucket(data_source),
  ...)
}
\arguments{
\item{x}{variable name}

\item{file}{path of the file to be written to}

\item{FUN}{the function using which the file is to write}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
No output
}
\description{
Write raster to GCS
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioWriteRaster(t, "tests/testras.tif", writeRaster, format = "GTiff")
}
}
