% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioWriteRda.R
\name{rioWriteRda}
\alias{rioWriteRda}
\title{Write RDA files}
\usage{
rioWriteRda(..., file, FUN = save, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source))
}
\arguments{
\item{...}{R pbjects need to be saved}

\item{file}{path of the file to be written to}

\item{FUN}{the function using which the file is to write}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}
}
\value{
No output
}
\description{
Write R data RDA file to anywhere from R
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioWriteRda(iris, mtcars, "tests/iris.rda")
}
}
