% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioWriteRds.R
\name{rioWriteRds}
\alias{rioWriteRds}
\title{Write RDS files}
\usage{
rioWriteRds(x, file, FUN = saveRDS, data_source = rioGetDataSource(),
  bucket = rioGetBucket(data_source), ...)
}
\arguments{
\item{x}{variable name}

\item{file}{path of the file to be written to}

\item{FUN}{the function using which the file is to write}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
if FUN returns anything
}
\description{
Write R data RDS file to anywhere from R
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioWriteRds(iris, "tests/iris.rds", saveRDS)
}
}
