% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioWriteShp.R
\name{rioWriteShp}
\alias{rioWriteShp}
\title{Write shapefiles}
\usage{
rioWriteShp(obj, dsn, layer, FUN = rgdal::writeOGR, dsnlayerbind = F,
  data_source = rioGetDataSource(), bucket = rioGetBucket(data_source),
  ...)
}
\arguments{
\item{obj}{R object to be written}

\item{dsn}{path of the file to be read}

\item{layer}{the name of the shapefile without extension}

\item{FUN}{the function using which the file is to be read}

\item{dsnlayerbind}{if the FUN needs dsn and layer binded or not}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
output of the FUN function if any
}
\description{
Write shapefiles
}
\examples{
\dontrun{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioWriteShp(t, "tests/shptest/", "new", driver = "ESRI Shapefile", overwrite = T)
}
}
