% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rioWriteTable.R
\name{rioWriteTable}
\alias{rioWriteTable}
\title{Write csv, Excel files, txt to GCS}
\usage{
rioWriteTable(x, file, FUN = write.csv,
  data_source = rioGetDataSource(), bucket = rioGetBucket(data_source),
  ...)
}
\arguments{
\item{x}{variable name}

\item{file}{path of the file to be written to}

\item{FUN}{the function using which the file is to write}

\item{data_source}{the name of the data source, if not set globally. s3, gcs or local}

\item{bucket}{the name of the bucket, if not set globally}

\item{...}{other parameters for the FUN function defined above}
}
\value{
No output
}
\description{
Write csv, Excel files, txt to GCS
}
\examples{
rioSetDataSource("gcs")
rioSetBucket("socialcops-test")
rioWriteTable(iris, "tests/iris.csv", write.csv)

}
