% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyio_datasource.R
\name{flyio_get_datasource}
\alias{flyio_get_datasource}
\title{Get global data source name for flyio}
\usage{
flyio_get_datasource()
}
\value{
the string - data source name stored
}
\description{
Get global data source name to be used for all the functions in flyio. Returns the value stored using flyio_set_datasource
}
\examples{
# first setting the data source
flyio_set_datasource("s3")
# getting the data source
flyio_get_datasource()
}
